==> src/r-processing/01_Code_volcano_selection.R <==
### project: Probabilistic ash fall hazard assessment to coral reefs in the Coral Triangle
### Jan-Christopher Fischer (jc.fischer@bristol.ac.uk)
### readme - input data

### AOI: Coral Triangle Initiative implementation zone including Singapore and Exclusive Economic Zones of Northern Mariana Islands and Guam
# Exclusive Economic Zones of Northern Mariana Islands and Guam were included because of good past and recent documentation of ash fall impacts on coral reef benthic and reef fish communities (Eldredge, 1982; Eldredge & Kropp; Vroom & Zgliczynski, 2011). 

# Coral Triangle Initiative (CTI) implementation zone:
# file "cti_cff_implementation_area.shp" was downloaded from http://ctatlas.coraltriangleinitiative.org/Dataset on 03.09.2021
# info: The Coral Triangle Initiative on Coral Reefs, Fisheries, and Food Security (CTI-CFF), multilateral partnership of Indonesia, Malaysia, Philippines, Papua New Guinea, Solomon Islands, Timor-Leste with the aim of sustainable use of natural marine and coastal resources. Key topics are food security, climate change, and marine biodiversity.

# EEZ data 
# file "eez_boundaries_v11.shp" Download from https://www.marineregions.org/downloads.php on 03.09.2021 
# Flanders Marine Institute (2019). Maritime Boundaries Geodatabase: Maritime Boundaries and Exclusive Economic Zones (200NM), version 11. Available online at https://www.marineregions.org/. https://doi.org/10.14284/386

# Processing in ArcMap: EEZ of Northern Mariana Islands, Guam and Singapore were combined with CTI zone
# Geoprocessing - Union
# Geoprocessing - Dissolve
# resultant file "AOI.shp" was saved
# 500 km buffer zone around AOI was generated by Geoprocessing - Buffer set linear unit to 500 km
# resultant file "AOI_buff_500_km.shp" was saved


### MPAs
# MPA data download for Guam, Indonesia, Malaysia, Northern Mariana Islands, Papua New Guinea, Philippines, Singapore, Solomon Islands, Timor-Leste from www.protectedplanet.net on 06.09.2021
# UNEP-WCMC and IUCN (2021), Protected Planet: The World Database on Protected Areas (WDPA) and World Database on Other Effective Area-based Conservation Measures (WD-OECM) [Online], September 2021, Cambridge, UK: UNEP-WCMC and IUCN. Available at: www.protectedplanet.net.

### land
# file "ne_10m_land.shp" was downloaded here file:///C:/Users/zs20225/OneDrive%20-%20University%20of%20Bristol/Documents/4D-REEF%20PhD%20Uni%20Bristol/Research/TephraProb%20modelling/Data/Original%20data/ne_10m_land/ne_10m_land.README.html on 14.04.2021

# Processing:
# out of 9 country-specific folders "WDPA_WDOECM_Aug2021_Public_COUNTRY_shp", all "WDPA_WDOECM_Aug2021_Public_COUNTRY_shp-polygons.shp" files were load in ArcMap
# Geoprocessing - Union
# Geoprocessing - Dissolve
# resultant file "WDPA_WDOECM_Sep2021_AOI_all.shp" was saved
# file "WDPA_WDOECM_Sep2021_AOI_all.shp" was cropped: in ArcMap: ArcToolbox - Analysis Tools - Overlay - Erase: Input Features were "WDPA_WDOECM_Sep2021_AOI_all.shp" and Erase Features were "ne_10m_land.shp"
# resultant file "WDPA_WDOECM_Sep2021_AOI_marine.shp" was saved

###########################################################################################################

### project: Probabilistic ash fall hazard assessment to coral reefs in the Coral Triangle
### Jan-Christopher Fischer (jc.fischer@bristol.ac.uk)
### select volcanoes to be included
### add coordinates for wind data download
### 08.09.2021

# check and clean memory
ls()
rm(list=ls(all.names=T))
gc()
ls()

# load libraries
library(raster)
library(rgeos)
library(rgdal)

# set working directory
setwd("$REPO_HOME/inputs/")

# load AOI shp
AOI <- shapefile("AOI.shp")
AOI
AOI_buff_500_km <- shapefile("AOI_buff_500_km.shp")
AOI_buff_500_km
extent(AOI_buff_500_km)
# xmin: 87.52166 
# xmax: 178.1652 
# ymin: -20.64415 
# ymax: 28.40833 

# load land data
land <- shapefile("ne_10m_land.shp")
land
land_crop <- crop(land, extent(AOI_buff_500_km))
land_crop
#plot(land_crop)

### load volcano data
#volc_holo <- read.csv("GVP_Volcano_List_Holocene_2020.csv", sep=";")
volc_holo <- read.csv("GVP_Volcano_List_Holocene.csv", sep=",")
# this file was downloaded here https://volcano.si.edu/list_volcano_holocene.cfm on 03.09.2021
volc_holo
head(volc_holo)
nrow(volc_holo)
names(volc_holo)

# exclude volcanoes by location (extent of AOI)
volc_holo <- volc_holo[volc_holo$Latitude  > -20.64415,]
volc_holo <- volc_holo[volc_holo$Latitude  < 28.40833 ,]
volc_holo <- volc_holo[volc_holo$Longitude > 87.52166 ,]
volc_holo <- volc_holo[volc_holo$Longitude < 178.1652 ,]
head(volc_holo)
nrow(volc_holo)

# exclude volcanoes by location (AOI)
# convert to SpatialPointsDataFrame
xy <- volc_holo[,c(10,9)]
xy
volc_holo_spat <- SpatialPointsDataFrame(coords=xy, data=volc_holo, proj4string=CRS("+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"))
volc_holo_spat
plot(volc_holo_spat)
volc_holo_spat_crop <- crop(volc_holo_spat, AOI_buff_500_km)
volc_holo_spat_crop
plot(volc_holo_spat_crop, col="red", add=T)
# convert back to data frame
volc_holo <- as.data.frame(volc_holo_spat_crop)
volc_holo
head(volc_holo)
nrow(volc_holo)

# exclude submarine volcanoes by type
volc_holo <- volc_holo[volc_holo$Primary.Volcano.Type != "Submarine",]
head(volc_holo)
nrow(volc_holo)

# exclude submarine volcanoes by elevation
volc_holo <- volc_holo[volc_holo$Elevation..m. > 0,]
head(volc_holo)
nrow(volc_holo)

# exclude volcanoes with uncertain evidence
volc_holo <- volc_holo[volc_holo$Activity.Evidence != "Evidence Uncertain",]
head(volc_holo)
nrow(volc_holo)

# list countries of retained volcanoes
sort(unique(volc_holo$Country))
### search at https://volcano.si.edu/search_eruption.cfm for eruptions from all those countries
# "Burma (Myanmar)"  "China"            "Fiji"             "India"            "Indonesia"        "Japan"           
# "Papua New Guinea" "Philippines"      "Solomon Islands"  "Taiwan"           "United States"    "Vanuatu"         
# "Vietnam"
# resultant file "GVP_Eruption_Results" was converted to csv

### load eruption data
eruptions <- read.csv("GVP_Eruption_Results.csv", sep=",")
eruptions
head(eruptions)
nrow(eruptions)
names(eruptions)
class(eruptions)

# only retain confirmed eruptions
unique(eruptions$Eruption.Category)
eruptions <- eruptions[eruptions$Eruption.Category == "Confirmed Eruption",]
head(eruptions)
nrow(eruptions)

# retain only eruption records of pre-selected volcanoes
eruptions_mody <- eruptions[eruptions$Volcano.Name %in% volc_holo$Volcano.Name,]
eruptions_mody
head(eruptions_mody)
nrow(eruptions_mody)

# check numbers of volcanoes
length(sort(unique(volc_holo$Volcano.Name)))
length(sort(unique(eruptions_mody$Volcano.Name)))

# retain only volcanoes with documented eruptions
volc_holo_mody <- volc_holo[volc_holo$Volcano.Name %in% eruptions_mody$Volcano.Name,]
volc_holo_mody
head(volc_holo_mody)
nrow(volc_holo_mody)

length(sort(unique(volc_holo_mody$Volcano.Name)))
sort(unique(volc_holo_mody$Volcano.Name))

#############
###### other option: only keep volcanoes with VEI>=2
###unique(eruptions_mody$VEI)
###eruptions_VEI <- eruptions_mody[eruptions_mody$VEI >= 2,]
###eruptions_VEI <- eruptions_VEI[!is.na(eruptions_VEI$VEI),]
###head(eruptions_VEI)
###nrow(eruptions_VEI)
###unique(eruptions_VEI$VEI)
###
#### retain only volcanoes with documented eruptions of VEI>=2
###volc_holo_VEI <- volc_holo_mody[volc_holo_mody$Volcano.Name %in% eruptions_VEI$Volcano.Name,]
###volc_holo_VEI
###head(volc_holo_VEI)
###nrow(volc_holo_VEI)
###
###length(sort(unique(volc_holo_VEI$Volcano.Name)))
###sort(unique(volc_holo_VEI$Volcano.Name))

##########
### exclude all columns but number, name, coordinates and elevation
volc_holo_mody <- volc_holo_mody[,c(1, 2, 10, 9, 11)]
head(volc_holo_mody)
###volc_holo_VEI <- volc_holo_VEI[,c(1, 2, 10, 9, 11)]
###head(volc_holo_VEI)

# sort alphabetically
volc_holo_mody <- volc_holo_mody[order(volc_holo_mody$Volcano.Name),]
head(volc_holo_mody)
###volc_holo_VEI <- volc_holo_VEI[order(volc_holo_VEI$Volcano.Name),]
###head(volc_holo_VEI)

### check tables
# volcanoes with documented eruptions
volc_holo_mody
head(volc_holo_mody)
nrow(volc_holo_mody)
length(sort(unique(volc_holo_mody$Volcano.Name)))
sort(unique(volc_holo_mody$Volcano.Name))

#### volcanoes with documented eruptions of VEI>=2
###volc_holo_VEI
###head(volc_holo_VEI)
###nrow(volc_holo_VEI)
###length(sort(unique(volc_holo_VEI$Volcano.Name)))
###sort(unique(volc_holo_VEI$Volcano.Name))

### add coordinates for wind data download
# add empty columns to be filled in loop
volc_holo_mody["wind_north"] <- NA
volc_holo_mody["wind_south"] <- NA
volc_holo_mody["wind_east" ] <- NA
volc_holo_mody["wind_west" ] <- NA
head(volc_holo_mody)

###volc_holo_VEI["wind_north"] <- NA
###volc_holo_VEI["wind_south"] <- NA
###volc_holo_VEI["wind_east" ] <- NA
###volc_holo_VEI["wind_west" ] <- NA
###head(volc_holo_VEI)

# rename tables
tab <- volc_holo_mody
###tab <- volc_holo_VEI
head(tab)

### loop over all volcanoes to add coordinates for wind data download
#i=1
for(i in 1:nrow(tab)){
  ### extract individual volcano coordinates
  lon <- tab[i,3]
  lon
  lat <- tab[i,4]
  lat
  
  ### identify north, south, east and west coordinates for wind points
  # North
  tab[i,6] <- round((lat + 0.01), 2)
  # South
  tab[i,7] <- round((lat - 0.01), 2)
  # East
  tab[i,8] <- round((lon + 0.01), 2)
  # West
  tab[i,9] <- round((lon - 0.01), 2)
  
  # check
  #head(tab)
  
  # print for progress monitoring
  print(paste0("finished ", i, " out of ", nrow(tab), " which is ", tab[i,1]))
}

# check
head(tab)

# write out csv
write.csv(tab, "volc_holo_mody.csv")
###write.csv(tab, "volc_holo_VEI.csv")

### done
