==> src/r-processing/05_Preparation_eruption_prob_VEI.R <==
### project: Probabilistic ash fall hazard assessment to coral reefs in the Coral Triangle
### Jan-Christopher Fischer (jc.fischer@bristol.ac.uk)
### create table with grid parameters for Tephra2/TephraProb modelling
### 26.11.2021

# check and clean memory
ls()
rm(list=ls(all.names=T))
gc()
ls()

# load libraries

# load volcano data (volcanoes selected for our study)
setwd("~")
tab_volc <- read.csv("volcano_set_wind.csv")
tab_volc
head(tab_volc)
class(tab_volc)

# load eruption probability data by Jenkins et al. in Loughlin et al. 2015
setwd("~")
tab_erupt_SJ <- read.csv("Book1.csv") # this table was generated by converting the original file "Volcano_Probs_from GVM-2013work_F-M.xlsx" to csv, unnecessary columns were excluded and table was copied into new excel sheet as is didn't open in R otherwise
tab_erupt_SJ
head(tab_erupt_SJ)
class(tab_erupt_SJ)

### modify volcano table
tab_input <- tab_volc
# excluded unnecessary columns
tab_input <- tab_input[c(2,3)]
head(tab_input)
# add columns
tab_input$VEI2_annual_prob <- NA
tab_input$VEI3_annual_prob <- NA
tab_input$VEI4_annual_prob <- NA
tab_input$VEI5_annual_prob <- NA
tab_input$VEI6_annual_prob <- NA
tab_input$VEI7_annual_prob <- NA
# check
head(tab_input)

### loop over all volcanoes
#i=1
for(i in 1:nrow(tab_input)){
  volc_number <- tab_input[i,1]
  volc_number
  
  ### process eruption probability data by Jenkins et al. in Loughlin et al. 2015 calculated by Susanna
  # check for this volcano number in probability table
  this_row_SJ <- tab_erupt_SJ[tab_erupt_SJ$Volcano.ID==volc_number,]
  this_row_SJ
  class(this_row_SJ)
  nrow(this_row_SJ)
  
  ### if-condition for available probability data per volcano
  if(nrow(this_row_SJ)==1){
    # write probabilities in volcano table
    tab_input[i,3] <- this_row_SJ[5][[1]]
    tab_input[i,4] <- this_row_SJ[6][[1]]
    tab_input[i,5] <- this_row_SJ[7][[1]]
    tab_input[i,6] <- this_row_SJ[8][[1]]
    tab_input[i,7] <- this_row_SJ[9][[1]]
    tab_input[i,8] <- this_row_SJ[10][[1]]
  } ### end if-condition for available probability data per volcano
  
  ### if-condition for unavailable probability data per volcano
  if(nrow(this_row_SJ)==0){
    # write probabilities in volcano table
    tab_input[i,3] <- NA
    tab_input[i,4] <- NA
    tab_input[i,5] <- NA
    tab_input[i,6] <- NA
    tab_input[i,7] <- NA
    tab_input[i,8] <- NA
  } ### end if-condition for unavailable probability data per volcano
  
  print(paste0("finished volcano ", i, " out of ", nrow(tab_input)))
} ### end loop over all volcanoes

# check
tab_input
head(tab_input)
nrow(tab_input)
ncol(tab_input)

# write out csv
setwd("~")
write.csv(tab_input, "volcano_set_erup.csv")

### done
